d2sstats: optimalize data size (fix #22)

+++ api.go
@@ -3,5 +3,5 @@ package d2d2s
 // SetLevel sets character's level
 func (d *D2S) SetLevel(level byte) {
 	d.Level = level
-	d.Stats.Level = uint64(level)
+	d.Stats.Level = uint32(level)
 }
